Add an implementation of Foldable.foldl for Persistent.Vector
This traversal is considerably faster than foldr and is faster than
traversing a Data.Sequence.  Traversing lists in either direction is
still faster.

I'm not sure if the strictness properties of this implementation are
quite right by the definition of Foldable.  Unfortunately,
Foldable.foldl' is implemented in terms of foldr, so it is stuck being
slow.  A directly exported version might be called for so that we can
provide good efficiency.