Many missing functions (following the Sequence API), slicing support
This patch adds O(1) slices of persistent vectors, along with many
derived functions.  The biggest change here is that the
traversals (foldr, foldl') had to be rewritten and are a bit more
complicated for the sliced case.  There are faster implementations
that are used for unsliced vectors, which will hopefully be the common
case.