Version bump
This version fixes a significant bug in the indexing functions.
The problem arose when a slice invalidated all of the elements
in the tree portion, leaving the only valid data in the tail.

Before, the tail wasn't updated properly.  Now it is and, in this
special case, the reference to the tree body is dropped.