A more thorough fix for the same bug as the last commit
This is more general and also fixes the handling of the same issue in
the update code.  The last fix only handled the case where the last
array in the tree was full.  The real problem is if a slice ends at
the end of any array in the tree.